Moved Publisher tests to Request test class and added specific Publisher tests - most of the tests writen were actually Request tests so they are moved to RequestTest - added tests for Publisher method - created a BaseTestCase with shered functionality - added `test/` to composer psr-4 autoload
